[python] Refactoring python test system
